requirements.txt
requests==2.26.0
websocket-client==1.2.1

